% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{datalong}
\alias{datalong}
\title{Bladder cancer data from Andrews & Herzberg}
\format{An object of class \code{data.frame} with 837 rows and 8 columns.}
\usage{
data(datalong)
}
\description{
Data from a randomized controlled trial of thiotepa vs. placebo for superficial bladder tumours
}
\references{
Byar D, Blackard C. Comparisons of placebo, pyridoxine, and topical thiotepa in preventing recurrence state in bladder cancer. Urology 1977; 10:556-561.
}
\keyword{datasets}
